requirements.txt
moviepy==2.1.1
numpy>=1.24.0,<2.0.0
Pillow>=11.1.0
pillow-heif>=0.16.0
